requirements.txt
annotated-types==0.6.0
asttokens==2.4.0
beartype==0.16.3
certifi==2023.7.22
charset-normalizer==3.3.0
cloudpickle==3.0.0
filelock==3.12.4
fsspec==2023.9.2
huggingface-hub==0.17.3
icontract==2.6.4
idna==3.4
interegular==0.3.2
Jinja2==3.1.2
lark==1.1.7
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2
numpy==1.26.1
outlines==0.0.8
packaging==23.2
perscache==0.6.1
Pillow==10.1.0
pydantic==2.4.2
pydantic_core==2.10.1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
scipy==1.11.3
six==1.16.0
sympy==1.12
tenacity==8.2.3
tokenizers==0.14.1
torch==2.1.0
tqdm==4.66.1
transformers==4.34.1
typing_extensions==4.8.0
urllib3==2.0.7

openai~=0.28.1
python-dotenv~=1.0.0
flask~=3.0.0
dacite~=1.8.1
pip~=23.2.1
attrs~=23.1.0
setuptools~=68.1.2
aiohttp~=3.8.5
yarl~=1.9.2